Separate components and pages (#24)
* Separate components and pages

* Improve style for rebased journey page
