+++ lib/model/inkstep_model.dart
@@ -1,0 +1,1 @@
+class InkstepModel {}
