+++ lib/page/onboarding.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:inkstep/component/default_flow_button.dart';
+import 'package:inkstep/component/user_flow_button.dart';
 import 'package:inkstep/logo.dart';
+import 'package:inkstep/main.dart';
+import 'package:inkstep/page/info_gathering.dart';
 import 'package:inkstep/theming.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-
-import 'info_gathering.dart';
-import 'main.dart';
-import 'onboarding/default_flow_button.dart';
-import 'onboarding/user_flow_button.dart';
 
 class Onboarding extends StatefulWidget {
   @override
@@ -69,7 +68,7 @@ class _OnboardingState extends State<Onboarding> with TickerProviderStateMixin {
     final bottom = Container(
       child: Column(
         children: <Widget>[
-          DefaultFlowButton(destination: NewJourneyRoute()),
+          BoldCallToAction(destination: InfoScreen()),
           Padding(padding: EdgeInsets.only(top: 32.0)),
           UserFlowButton(destination: TopTabs()),
         ],
