+++ lib/page/studios.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-
-import 'journeys/journeys.dart';
+import 'package:inkstep/page/journey.dart';
 
 class StudiosPage extends StatelessWidget {
   @override
